first attempts of a toggle alpha button
The toggle alpha button isn't working. But it's there. The plan is to
toggle alpha on every set_color tool, so that i.e. it's possible to
make an alpha gradient over an existing image without the need to add
an alpha gradient tool (cause there's already a gradient tool).
Next to the button should be a slider and a text-box to set alpha
form 0 to 1 or... let's say, from 0 to 100 which will be divided by 100